+++ src/main/java/com/HIPAA/SeniorProject/Repository/RolesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Role, Integer> {
+public interface RolesRepository extends JpaRepository<Role, Long> {
 }
