+++ src/main/java/com/HIPAA/SeniorProject/Repository/UserCredentialsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserCredentialsRepository extends JpaRepository<UserCredentials, Integer>  {
+public interface UserCredentialsRepository extends JpaRepository<UserCredentials, Long>  {
 
     @Query("SELECT c.password FROM UserCredentials c LEFT JOIN User s ON c.user = s.id WHERE s.email = ?1")
     String findUsersPassword(String email);
